delete account asks confirmation pop up

+++ frontend/src/routes/profile/+page.svelte
@@ -16,33 +16,36 @@
 
   async function deleteUser() {
 
-let user_id = data.user.user_id;
-let is_admin = data.user.is_admin;
+    let confirmDelAcc = confirm("Are you sure you want to delete your account?");
+    if (confirmDelAcc) {
 
-try {
-  const response = await fetch(`${USER_URL}/${user_id}`, {
-    method: 'DELETE',
-    headers: { "Content-Type": "application/json" },
-    body: JSON.stringify({ user_id , is_admin })
-  });
+    let user_id = data.user.user_id;
+    let is_admin = data.user.is_admin;
 
-  if (response.ok) {
-    // Redirect to home page or another page after successful deletion
-    goto('/');
+    try {
+      const response = await fetch(`${USER_URL}/${user_id}`, {
+        method: "DELETE",
+        headers: { "Content-Type": "application/json" },
+        body: JSON.stringify({ user_id, is_admin })
+      });
 
-    const response = await fetch(LOGOUT_URL, {
-      method: "POST",
-      credentials: "include"
-    });
-    await invalidateAll();
-  
-  } else {
-    console.error('Failed to delete user:', response.statusText);
+      if (response.ok) {
+        // Redirect to home page or another page after successful deletion
+        goto("/");
+
+        const response = await fetch(LOGOUT_URL, {
+          method: "POST",
+          credentials: "include"
+        });
+        await invalidateAll();
+      } else {
+        console.error("Failed to delete user:", response.statusText);
+      }
+    } catch (error) {
+      console.error("Error deleting user:", error);
+    }
   }
-} catch (error) {
-  console.error('Error deleting user:', error);
-}
-}
+  }
 </script>
 
 <svelte:head>
@@ -51,79 +54,79 @@ try {
 
 {#if data.user}
   {#if editMode}
-    <UserInfoForm user={data.user} on:save={toggleEditMode} on:cancel={toggleEditMode}/>
+    <UserInfoForm user={data.user} on:save={toggleEditMode} on:cancel={toggleEditMode} />
   {:else}
     <div class="main">
       <h2>My Profile</h2>
       <div class="card">
-              <table>
-                  <tbody>
-                      <tr>
-                          <td>First Name</td>
-                          <td>:</td>
-                          <td>{data.user.firstName}</td>
-                      </tr>
-                      <tr>
-                          <td>Last Name</td>
-                          <td>:</td>
-                          <td>{data.user.lastName}</td>
-                      </tr>
-                      <tr>
-                          <td>Date of birth</td>
-                          <td>:</td>
-                          <td>{data.user.dob}</td>
-                      </tr>
-                      <tr>
-                          <td>Username</td>
-                          <td>:</td>
-                          <td>{data.user.username}</td>
-                      </tr>
-                      <tr>
-                          <td>Description</td>
-                          <td>:</td>
-                          <td>{data.user.desc}</td>
-                      </tr>
-                      <tr>
-                          <td>Avatar</td>
-                          <td>:</td>
-                          <td>{data.user.avatar}</td>
-                      </tr>
-                  </tbody>
-              </table>
-              <button on:click={toggleEditMode}>Edit</button>
-              <button on:click={deleteUser}>Delete Account</button>
-          </div>
+        <table>
+          <tbody>
+            <tr>
+              <td>First Name</td>
+              <td>:</td>
+              <td>{data.user.firstName}</td>
+            </tr>
+            <tr>
+              <td>Last Name</td>
+              <td>:</td>
+              <td>{data.user.lastName}</td>
+            </tr>
+            <tr>
+              <td>Date of birth</td>
+              <td>:</td>
+              <td>{data.user.dob}</td>
+            </tr>
+            <tr>
+              <td>Username</td>
+              <td>:</td>
+              <td>{data.user.username}</td>
+            </tr>
+            <tr>
+              <td>Description</td>
+              <td>:</td>
+              <td>{data.user.desc}</td>
+            </tr>
+            <tr>
+              <td>Avatar</td>
+              <td>:</td>
+              <td>{data.user.avatar}</td>
+            </tr>
+          </tbody>
+        </table>
+        <button on:click={toggleEditMode}>Edit</button>
+        <button on:click={deleteUser}>Delete Account</button>
       </div>
+    </div>
   {/if}
 {/if}
-<style>
 
-.main {
+<style>
+  .main {
     margin: auto;
     font-size: 28px;
     padding: 50px;
     width: 50%;
-}
+  }
 
-.main h2 {
+  .main h2 {
     color: #333;
     font-size: 30px;
     margin-bottom: 20px;
-}
+  }
 
-.main .card {
+  .main .card {
     background-color: #fff;
     border-radius: 18px;
     box-shadow: 1px 1px 8px 0 grey;
     height: auto;
     margin-bottom: 20px;
     padding: 20px 0 20px 50px;
-}
+  }
 
-.main .card table {
+  .main .card table {
     border: none;
     font-size: 16px;
     height: 270px;
     width: 80%;
-}
+  }
 </style>
